safe_request to utils
